Don't patch DOM nodes if the element's text equals new node.
When working with text nodes and content-editable, it is possible
we end up patching a text node where element.nodeValue is the same
as the new node text. In this situation, the state is already sync
with the DOM.

If we ignore this and replace the old DOM, the content-editable
element will run into focus / selection reset issues.