Remove modules. (#453)
Modules are no doubt a convenient mechanism to "safely" inject state and action slices into your model. The problem is now we have two ways to define our state and actions.

This means it is very hard to wrap your state and actions outside the app() call, e.g, to implement your own module assembly functionality.

    import Global from "./global"
    import Nav from "./nav"
    import Layout from "./layout"
    import module from "./module"

    app(
      module({
        Global, Nav, Layout
      })
    )

Without modules, we can assemble the state outside the app call if we need to. This way it's up to you how you mix the state with other sources.

    const state = JSON.parse(localStorage.getItem("state"))

    app({
      state,
      view,
      actions
    })

Instead of removing modules, one could fix the problem above if we bring back init. Init would be called with the state and actions after they have been assembled/wired by Hyperapp. I want to avoid introducing more complexity, so I really want to move forward with this PR instead.

Removing modules means you will need to explicitly assemble your state and actions (and possibly subscriptions in the future). Explicit is good, don't be afraid of it.

  import foo from "./foo"
  import bar from "./bar"

  app({
    state: {
      foo: foo.state,
      bar: bar.state
    },
    actions: {
      foo: foo.actions,
      bar: bar.actions
    }
  })